Move all '*Payload' classes inside their API counterparts
Based on review feedback from Mike Drob on PR #565
