+++ solr/core/src/java/org/apache/solr/handler/admin/api/DeleteReplicaPropertyAPI.java
@@ -20,8 +20,9 @@ package org.apache.solr.handler.admin.api;
 import org.apache.solr.api.Command;
 import org.apache.solr.api.EndPoint;
 import org.apache.solr.api.PayloadObj;
-import org.apache.solr.client.solrj.request.beans.DeleteReplicaPropertyPayload;
+import org.apache.solr.common.annotation.JsonProperty;
 import org.apache.solr.common.params.CollectionParams;
+import org.apache.solr.common.util.ReflectMapWriter;
 import org.apache.solr.handler.admin.CollectionsHandler;
 
 import java.util.HashMap;
@@ -62,5 +63,16 @@ public class DeleteReplicaPropertyAPI {
     v1Params.put(COLLECTION, obj.getRequest().getPathTemplateValues().get(COLLECTION));
 
     collectionsHandler.handleRequestBody(wrapParams(obj.getRequest(), v1Params), obj.getResponse());
+  }
+
+  public static class DeleteReplicaPropertyPayload implements ReflectMapWriter {
+    @JsonProperty(required = true)
+    public String shard;
+
+    @JsonProperty(required = true)
+    public String replica;
+
+    @JsonProperty(required = true)
+    public String property;
   }
 }
