+++ solr/core/src/java/org/apache/solr/handler/admin/api/SetCollectionPropertyAPI.java
@@ -20,8 +20,9 @@ package org.apache.solr.handler.admin.api;
 import org.apache.solr.api.Command;
 import org.apache.solr.api.EndPoint;
 import org.apache.solr.api.PayloadObj;
-import org.apache.solr.client.solrj.request.beans.SetCollectionPropertyPayload;
+import org.apache.solr.common.annotation.JsonProperty;
 import org.apache.solr.common.params.CollectionParams;
+import org.apache.solr.common.util.ReflectMapWriter;
 import org.apache.solr.handler.admin.CollectionsHandler;
 
 import java.util.HashMap;
@@ -39,8 +40,6 @@ import static org.apache.solr.security.PermissionNameProvider.Name.COLL_EDIT_PER
  *
  * This API (POST /v2/collections/collectionName {'set-collection-property': {...}}) is analogous to the v1
  * /admin/collections?action=COLLECTIONPROP command.
- *
- * @see SetCollectionPropertyPayload
  */
 @EndPoint(
         path = {"/c/{collection}", "/collections/{collection}"},
@@ -68,5 +67,13 @@ public class SetCollectionPropertyAPI {
     v1Params.put(NAME, obj.getRequest().getPathTemplateValues().get(COLLECTION));
 
     collectionsHandler.handleRequestBody(wrapParams(obj.getRequest(), v1Params), obj.getResponse());
+  }
+
+  public static class SetCollectionPropertyPayload implements ReflectMapWriter {
+    @JsonProperty(required = true)
+    public String name;
+
+    @JsonProperty
+    public String value = null;
   }
 }
