--- solr/solrj/src/java/org/apache/solr/client/solrj/request/beans/MigrateDocsPayload.java
@@ -1,37 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.solr.client.solrj.request.beans;
-
-import org.apache.solr.common.annotation.JsonProperty;
-import org.apache.solr.common.util.ReflectMapWriter;
-
-public class MigrateDocsPayload implements ReflectMapWriter {
-  @JsonProperty(required = true)
-  public String target;
-
-  @JsonProperty(required = true)
-  public String splitKey;
-
-  @JsonProperty
-  public Integer forwardTimeout = 60;
-
-  @JsonProperty
-  public Boolean followAliases;
-
-  @JsonProperty
-  public String async;
-}
